+++ collector/config.ini
@@ -7,9 +7,9 @@ url = http://api.openweathermap.org/data/2.5/weather
 api_key = 51f9d590eeeb790b4e8e2d218307f523
 
 [db]
-sa_url = postgresql://myuser:mypass@172.19.0.3/mydb
-test_sa_url = postgresql://myuser:mypass@172.19.0.3/mydb_test
-alembic_sa_url = postgresql://myuser:mypass@172.19.0.3/mydb
+sa_url = postgresql://myuser:mypass@172.25.0.2/mydb
+test_sa_url = postgresql://myuser:mypass@172.25.0.2/mydb_test
+alembic_sa_url = postgresql://myuser:mypass@172.25.0.2/mydb
 sa_echo = 0
 table = city_weather
