+++ collector/config.ini.template
