+++ docker-compose.yml
@@ -4,9 +4,6 @@ services:
     build:
       context: .
       dockerfile: Dockerfile.test
-    volumes:
-      - ./config.ini:/app/config.ini
-      - ./cities.py:/app/cities.py
     command: sh -c "./wait-for-it.sh db:${DB_PORT:-5432} && python collector.py"
     depends_on:
       - db
@@ -24,16 +21,5 @@ services:
       POSTGRES_PASSWORD: ${DB_PASSWORD}
       DB_PORT: ${DB_PORT:-5432}
     restart: unless-stopped
-  redis:
-    image: redis:7.0.0-alpine
-    ports:
-      - ${REDIS_PORT:-6379}:6379
-    volumes:
-      - /opt/redis/etc/redis.conf:/usr/local/etc/redis/redis.conf
-      - /opt/redis/data:/data
-    restart: unless-stopped
-    command: /usr/local/etc/redis/redis.conf
 volumes:
   database: {}
-  redis: {}
-
